Move to better ClientEvents
