+++ src/main/java/com/wynntils/modules/utilities/events/ClientEvents.java
@@ -7,7 +7,6 @@ package com.wynntils.modules.utilities.events;
 import com.wynntils.ModCore;
 import com.wynntils.Reference;
 import com.wynntils.core.events.custom.*;
-import com.wynntils.core.framework.FrameworkManager;
 import com.wynntils.core.framework.enums.wynntils.WynntilsSound;
 import com.wynntils.core.framework.instances.PlayerInfo;
 import com.wynntils.core.framework.interfaces.Listener;
@@ -31,7 +30,6 @@ import net.minecraft.client.entity.EntityPlayerSP;
 import net.minecraft.client.gui.GuiScreen;
 import net.minecraft.client.gui.GuiYesNo;
 import net.minecraft.entity.Entity;
-import net.minecraft.entity.item.EntityArmorStand;
 import net.minecraft.entity.passive.AbstractHorse;
 import net.minecraft.entity.player.InventoryPlayer;
 import net.minecraft.init.SoundEvents;
@@ -41,7 +39,9 @@ import net.minecraft.inventory.InventoryBasic;
 import net.minecraft.item.ItemStack;
 import net.minecraft.network.play.client.*;
 import net.minecraft.network.play.client.CPacketPlayerDigging.Action;
-import net.minecraft.network.play.server.*;
+import net.minecraft.network.play.server.SPacketEntityMetadata;
+import net.minecraft.network.play.server.SPacketSetSlot;
+import net.minecraft.network.play.server.SPacketWindowItems;
 import net.minecraft.scoreboard.Team;
 import net.minecraft.util.EnumHand;
 import net.minecraft.util.text.ChatType;
@@ -49,19 +49,14 @@ import net.minecraft.util.text.TextComponentString;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraftforge.client.event.*;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
-import net.minecraftforge.fml.common.eventhandler.Event;
 import net.minecraftforge.fml.common.eventhandler.EventPriority;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.common.gameevent.InputEvent;
 import net.minecraftforge.fml.common.gameevent.TickEvent;
 import org.lwjgl.opengl.Display;
 
-import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.stream.IntStream;
 
 public class ClientEvents implements Listener {
     private static GuiScreen scheduledGuiScreen = null;
@@ -282,198 +277,6 @@ public class ClientEvents implements Listener {
         if (msg.startsWith("[Daily Rewards:")) {
             DailyReminderManager.openedDaily();
         }
-    }
-
-    TotemTracker totemTracker = new TotemTracker();
-
-    private static class TotemTracker {
-        public enum TotemState { NONE, SUMMONED, LANDING, PREPARING, ACTIVATING, ACTIVE}
-        private TotemState totemState = TotemState.NONE;
-        private int trackedTotemId = -1;
-        private double trackedX, trackedY, trackedZ;
-        private double potentialX, potentialY, potentialZ;
-        private int potentialTrackedId = -1;
-        private int trackedTime;
-        private long spellCastTimestamp = 0;
-        private long totemCreatedTimestamp = Long.MAX_VALUE;
-
-        private int bufferedId = -1;
-        private double bufferedX = -1;
-        private double bufferedY = -1;
-        private double bufferedZ = -1;
-
-        private void updateTotemPosition(double x, double y, double z) {
-            trackedX = x;
-            trackedY = y;
-            trackedZ = z;
-        }
-
-        private void resetTotemTracking() {
-            trackedTotemId = -1;
-            trackedTime = -1;
-            trackedX = 0;
-            trackedY = 0;
-            trackedZ = 0;
-        }
-
-        private static boolean isClose(double a, double b)
-        {
-            double diff = Math.abs(a - b);
-            return  (diff < 3);
-        }
-
-        private Entity getBufferedEntity(int entityId) {
-            Entity entity = ModCore.mc().world.getEntityByID(entityId);
-            if (entity != null) return entity;
-
-            if (entityId == bufferedId) {
-                return new EntityArmorStand(ModCore.mc().world, bufferedX, bufferedY, bufferedZ);
-            }
-
-            return null;
-        }
-
-        private void postEvent(Event event) {
-            ModCore.mc().addScheduledTask(() -> FrameworkManager.getEventBus().post(event));
-        }
-
-        private void checkTotemSummoned() {
-            // Check if we have both creation and spell cast at roughly the same time
-            if (Math.abs(totemCreatedTimestamp - spellCastTimestamp) < 500) {
-                if (totemState != TotemState.NONE) {
-                    // We have an active totem already, first remove that one
-                    totemState = TotemState.NONE;
-                    resetTotemTracking();
-                    postEvent(new SpellEvent.TotemRemoved(true));
-                }
-                trackedTotemId = potentialTrackedId;
-                trackedTime  = -1;
-
-                updateTotemPosition(potentialX, potentialY, potentialZ);
-                totemState = TotemState.SUMMONED;
-                postEvent(new SpellEvent.TotemSummoned());
-            }
-        }
-
-        public void onTotemSpawn(PacketEvent<SPacketSpawnObject> e) {
-            if (e.getPacket().getType() == 78) {
-                bufferedId = e.getPacket().getEntityID();
-                bufferedX = e.getPacket().getX();
-                bufferedY = e.getPacket().getY();
-                bufferedZ = e.getPacket().getZ();
-
-                if (e.getPacket().getEntityID() == trackedTotemId) {
-                    // Totems respawn with the same entityID when landing.
-                    // Update with more precise coordinates
-                    updateTotemPosition(e.getPacket().getX(), e.getPacket().getY(), e.getPacket().getZ());
-                    totemState = TotemState.LANDING;
-                    return;
-                }
-
-                // Is it created close to us? Then it's a potential new totem
-                if (isClose(e.getPacket().getX(), Minecraft.getMinecraft().player.posX) &&
-                        isClose(e.getPacket().getY(), Minecraft.getMinecraft().player.posY + 1.0) &&
-                        isClose(e.getPacket().getZ(), Minecraft.getMinecraft().player.posZ)) {
-                    potentialTrackedId = e.getPacket().getEntityID();
-                    potentialX = e.getPacket().getX();
-                    potentialY = e.getPacket().getY();
-                    potentialZ = e.getPacket().getZ();
-                    totemCreatedTimestamp = System.currentTimeMillis();
-                    checkTotemSummoned();
-                }
-            }
-        }
-
-        public void onTotemSpellCast(SpellEvent.Cast e) {
-            if (e.getSpell().equals("Totem")) {
-                spellCastTimestamp = System.currentTimeMillis();
-                checkTotemSummoned();
-            }
-        }
-
-        public void onTotemTeleport(PacketEvent<SPacketEntityTeleport> e) {
-            int thisId = e.getPacket().getEntityId();
-
-            if (thisId == trackedTotemId && (totemState == TotemState.SUMMONED || totemState == TotemState.LANDING)) {
-                // Now the totem has gotten it's final coordinates
-                updateTotemPosition(e.getPacket().getX(), e.getPacket().getY(), e.getPacket().getZ());
-                totemState = TotemState.PREPARING;
-            }
-        }
-
-        public void onTotemRename(PacketEvent<SPacketEntityMetadata> e) {
-            if (!Reference.onServer || !Reference.onWorld) return;
-
-            String name = Utils.getNameFromMetadata(e.getPacket().getDataManagerEntries());
-            if (name == null || name.isEmpty()) return;
-
-            Entity entity = getBufferedEntity(e.getPacket().getEntityId());
-            if (!(entity instanceof EntityArmorStand)) return;
-
-            Pattern shamanTotemTimer = Pattern.compile("^§c([0-9][0-9]?)s$");
-            Matcher m = shamanTotemTimer.matcher(name);
-            if (m.find()) {
-                // We got a armor stand with a timer nametag
-                double distanceXZ = Math.abs(entity.posX  - trackedX) +  Math.abs(entity.posZ  - trackedZ);
-                if (distanceXZ < 3.0 && entity.posY <= (trackedY + 3.0) && entity.posY >= ((trackedY + 2.0))) {
-                    // ... and it's close to our totem; regard this as our timer
-                    int time = Integer.parseInt(m.group(1));
-
-                    if (trackedTime == -1) {
-                        trackedTime = time;
-                        totemState = TotemState.ACTIVE;
-                        postEvent(new SpellEvent.TotemActivated(trackedTime));
-                    } else if (time != trackedTime) {
-                        trackedTime = time;
-                    }
-                }
-            }
-        }
-
-        public void onTotemDestroy(PacketEvent<SPacketDestroyEntities> e) {
-            IntStream entityIDs = Arrays.stream(e.getPacket().getEntityIDs());
-            if (entityIDs.filter(id -> id == trackedTotemId).findFirst().isPresent()) {
-                if (totemState == TotemState.ACTIVE && trackedTime == 0) {
-                    totemState = TotemState.NONE;
-                    resetTotemTracking();
-                    postEvent(new SpellEvent.TotemRemoved(false));
-                }
-            }
-        }
-
-        public void onTotemClassChange(WynnClassChangeEvent e) {
-            resetTotemTracking();
-        }
-    }
-
-    @SubscribeEvent
-    public void onTotemSpawn(PacketEvent<SPacketSpawnObject> e) {
-        totemTracker.onTotemSpawn(e);
-    }
-
-    @SubscribeEvent
-    public void onTotemSpellCast(SpellEvent.Cast e) {
-        totemTracker.onTotemSpellCast(e);
-    }
-
-    @SubscribeEvent
-    public void onTotemTeleport(PacketEvent<SPacketEntityTeleport> e) {
-        totemTracker.onTotemTeleport(e);
-    }
-
-    @SubscribeEvent
-    public void onTotemRename(PacketEvent<SPacketEntityMetadata> e) {
-        totemTracker.onTotemRename(e);
-    }
-
-    @SubscribeEvent
-    public void onTotemDestroy(PacketEvent<SPacketDestroyEntities> e) {
-        totemTracker.onTotemDestroy(e);
-    }
-
-    @SubscribeEvent
-    public void onTotemClassChange(WynnClassChangeEvent e) {
-        totemTracker.onTotemClassChange(e);
     }
 
     @SubscribeEvent
